fix: remove unused hook, move hook to module
